* Add simple test scripts to verify that the configs are created * Add an option to gke/deploy.sh to skip deployment to facilitate testing * Add an option to gke/deploy.sh to not recreate the DM configs if they   already exist; the sed commands assume we are starting from scratch.

+++ scripts/gke/deploy.sh
@@ -11,6 +11,7 @@ set -xe
 
 KUBEFLOW_REPO=${KUBEFLOW_REPO:-"`pwd`/kubeflow_repo"}
 KUBEFLOW_VERSION=${KUBEFLOW_VERSION:-"master"}
+KUBEFLOW_DEPLOY=${KUBEFLOW_DEPLOY:-true}
 
 if [[ ! -d "${KUBEFLOW_REPO}" ]]; then
   if [ "${KUBEFLOW_VERSION}" == "master" ]; then
@@ -62,7 +63,11 @@ SETUP_PROJECT=${SETUP_PROJECT:true}
 # Namespace where kubeflow is deployed
 K8S_NAMESPACE=${K8S_NAMESPACE:-"kubeflow"}
 CONFIG_FILE=${CONFIG_FILE:-"cluster-kubeflow.yaml"}
-PROJECT_NUMBER=`gcloud projects describe ${PROJECT} --format='value(project_number)'`
+
+if [ -z "${PROJECT_NUMBER}" ]; then 
+  PROJECT_NUMBER=`gcloud projects describe ${PROJECT} --format='value(project_number)'`
+fi
+
 ADMIN_EMAIL=${DEPLOYMENT_NAME}-admin@${PROJECT}.iam.gserviceaccount.com
 USER_EMAIL=${DEPLOYMENT_NAME}-user@${PROJECT}.iam.gserviceaccount.com
 
@@ -82,48 +87,56 @@ else
   echo skipping project setup
 fi
 
-# Check if it already exists
-set +e
-gcloud deployment-manager --project=${PROJECT} deployments describe ${DEPLOYMENT_NAME}
-exists=$?
-set -e
-
-cp -r "${KUBEFLOW_REPO}/scripts/gke/deployment_manager_configs" "${KUBEFLOW_DM_DIR}"
-cd "${KUBEFLOW_DM_DIR}"
-# Set values in DM config file
-sed -i.bak "s/zone: us-central1-a/zone: ${ZONE}/" "${KUBEFLOW_DM_DIR}/${CONFIG_FILE}"
-sed -i.bak "s/users:/users: [\"user:${EMAIL}\"]/" "${KUBEFLOW_DM_DIR}/${CONFIG_FILE}"
-sed -i.bak "s/ipName: kubeflow-ip/ipName: ${KUBEFLOW_IP_NAME}/" "${KUBEFLOW_DM_DIR}/${CONFIG_FILE}"
-rm "${KUBEFLOW_DM_DIR}/${CONFIG_FILE}.bak"
-
-if [ ${exists} -eq 0 ]; then
-  echo ${DEPLOYMENT_NAME} exists
-  gcloud deployment-manager --project=${PROJECT} deployments update ${DEPLOYMENT_NAME} --config=${CONFIG_FILE}
+# Create the DM configs if they don't exists
+if [ ! -d "${KUBEFLOW_DM_DIR}" ]; then
+  echo creating Deployment Manager configs in directory "${KUBEFLOW_DM_DIR}"
+  cp -r "${KUBEFLOW_REPO}/scripts/gke/deployment_manager_configs" "${KUBEFLOW_DM_DIR}"
+  cd "${KUBEFLOW_DM_DIR}"
+  # Set values in DM config file
+  sed -i.bak "s/zone: us-central1-a/zone: ${ZONE}/" "${KUBEFLOW_DM_DIR}/${CONFIG_FILE}"
+  sed -i.bak "s/users:/users: [\"user:${EMAIL}\"]/" "${KUBEFLOW_DM_DIR}/${CONFIG_FILE}"
+  sed -i.bak "s/ipName: kubeflow-ip/ipName: ${KUBEFLOW_IP_NAME}/" "${KUBEFLOW_DM_DIR}/${CONFIG_FILE}"
+  rm "${KUBEFLOW_DM_DIR}/${CONFIG_FILE}.bak"
 else
-  # Run Deployment Manager
-  gcloud deployment-manager --project=${PROJECT} deployments create ${DEPLOYMENT_NAME} --config=${CONFIG_FILE}
+  echo Deployment Manager configs already exist in directory "${KUBEFLOW_DM_DIR}"
 fi
 
-# TODO(jlewi): We should name the secrets more consistently based on the service account name.
-# We will need to update the component configs though
-gcloud --project=${PROJECT} iam service-accounts keys create ${ADMIN_EMAIL}.json --iam-account ${ADMIN_EMAIL}
-gcloud --project=${PROJECT} iam service-accounts keys create ${USER_EMAIL}.json --iam-account ${USER_EMAIL}
+if ${KUBEFLOW_DEPLOY}; then
+  # Check if it already exists
+  set +e
+  gcloud deployment-manager --project=${PROJECT} deployments describe ${DEPLOYMENT_NAME}
+  exists=$?
+  set -e
 
-# Set credentials for kubectl context
-gcloud --project=${PROJECT} container clusters get-credentials --zone=${ZONE} ${DEPLOYMENT_NAME}
+  if [ ${exists} -eq 0 ]; then
+    echo ${DEPLOYMENT_NAME} exists
+    gcloud deployment-manager --project=${PROJECT} deployments update ${DEPLOYMENT_NAME} --config=${CONFIG_FILE}
+  else
+    # Run Deployment Manager
+    gcloud deployment-manager --project=${PROJECT} deployments create ${DEPLOYMENT_NAME} --config=${CONFIG_FILE}
+  fi
 
-# Make yourself cluster admin
-kubectl create clusterrolebinding default-admin --clusterrole=cluster-admin --user=${EMAIL}
+  # TODO(jlewi): We should name the secrets more consistently based on the service account name.
+  # We will need to update the component configs though
+  gcloud --project=${PROJECT} iam service-accounts keys create ${ADMIN_EMAIL}.json --iam-account ${ADMIN_EMAIL}
+  gcloud --project=${PROJECT} iam service-accounts keys create ${USER_EMAIL}.json --iam-account ${USER_EMAIL}
 
-kubectl create namespace ${K8S_NAMESPACE}
+  # Set credentials for kubectl context
+  gcloud --project=${PROJECT} container clusters get-credentials --zone=${ZONE} ${DEPLOYMENT_NAME}
 
-# We want the secret name to be the same by default for all clusters so that users don't have to set it manually.
-kubectl create secret generic --namespace=${K8S_NAMESPACE} admin-gcp-sa --from-file=admin-gcp-sa.json=./${ADMIN_EMAIL}.json
-kubectl create secret generic --namespace=${K8S_NAMESPACE} user-gcp-sa --from-file=user-gcp-sa.json=./${USER_EMAIL}.json
-kubectl create secret generic --namespace=${K8S_NAMESPACE} kubeflow-oauth --from-literal=CLIENT_ID=${CLIENT_ID} --from-literal=CLIENT_SECRET=${CLIENT_SECRET}
+  # Make yourself cluster admin
+  kubectl create clusterrolebinding default-admin --clusterrole=cluster-admin --user=${EMAIL}
 
-# Install the GPU driver. It has no effect on non-GPU nodes.
-kubectl apply -f https://raw.githubusercontent.com/GoogleCloudPlatform/container-engine-accelerators/stable/nvidia-driver-installer/cos/daemonset-preloaded.yaml
+  kubectl create namespace ${K8S_NAMESPACE}
+
+  # We want the secret name to be the same by default for all clusters so that users don't have to set it manually.
+  kubectl create secret generic --namespace=${K8S_NAMESPACE} admin-gcp-sa --from-file=admin-gcp-sa.json=./${ADMIN_EMAIL}.json
+  kubectl create secret generic --namespace=${K8S_NAMESPACE} user-gcp-sa --from-file=user-gcp-sa.json=./${USER_EMAIL}.json
+  kubectl create secret generic --namespace=${K8S_NAMESPACE} kubeflow-oauth --from-literal=CLIENT_ID=${CLIENT_ID} --from-literal=CLIENT_SECRET=${CLIENT_SECRET}
+
+  # Install the GPU driver. It has no effect on non-GPU nodes.
+  kubectl apply -f https://raw.githubusercontent.com/GoogleCloudPlatform/container-engine-accelerators/stable/nvidia-driver-installer/cos/daemonset-preloaded.yaml
+fi
 
 # Create the ksonnet app
 cd $(dirname "${KUBEFLOW_KS_DIR}")
@@ -156,7 +169,9 @@ ks param set kubeflow-core reportUsage true
 ks param set kubeflow-core usageId $(uuidgen)
 
 # Apply the components generated
-ks apply default -c kubeflow-core
-ks apply default -c cloud-endpoints
-ks apply default -c cert-manager
-ks apply default -c iap-ingress
+if ${KUBEFLOW_DEPLOY}; then
+  ks apply default -c kubeflow-core
+  ks apply default -c cloud-endpoints
+  ks apply default -c cert-manager
+  ks apply default -c iap-ingress
+fi
